Binary expression filled out. DebugWriter converted to use visitor pattern.

+++ Mika/FloatConstantExpression.h
@@ -5,7 +5,7 @@
 class FloatConstantExpression : public Expression
 {
 protected:
-	float mValue;
+	double mValue;
 
 public:
 	FloatConstantExpression(size_t rootToken)
